Added test gadget for testing inside an iframe
I had to change the karma basePath for this, as otherwise EventEmitter.js is served on an absolute path, which we don’t know in test_gadget.html.

We might want to think about the name of the repo, because if you install this repo with bower the directory name will be `versal-player-api`, and the tests won’t work, as they depend on the name being `player-api`.